job/requirements.txt
analytics-python
google-cloud-automl
gcsfs
pandas
stacklogging
